# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Docmaster API",
        "description": "Docmaster API Documentation",
        "contact": {
            "name": "LJN",
            "email": "jordanndjock07@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/declaration": {
            "get": {
                "tags": [
                    "Déclarations"
                ],
                "summary": "Lister toutes les déclarations",
                "description": "Nécessite un token Bearer dans l'en-tête Authorization",
                "operationId": "a101191f8b3d361acf612ea6e67656a2",
                "responses": {
                    "200": {
                        "description": "Liste paginée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "declarations": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Docmaster"
                                                    }
                                                },
                                                "meta": {
                                                    "type": "object",
                                                    "example": {
                                                        "total": 20,
                                                        "current_page": 1,
                                                        "per_page": 10
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Liste des déclarations récupérée avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Déclarations"
                ],
                "summary": "Déclarer un document trouvé ou perdu",
                "description": "Nécessite un token Bearer dans l'en-tête Authorization",
                "operationId": "8d3eb4e28eaad378cd309aec76042699",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "type_docmaster": {
                                        "type": "string",
                                        "example": "Chercher | Trouver"
                                    },
                                    "date_action": {
                                        "type": "string",
                                        "example": "2024-09-23"
                                    },
                                    "type_document_id": {
                                        "type": "string",
                                        "example": "uuid-type-doc"
                                    },
                                    "nom_proprietaire": {
                                        "type": "string",
                                        "example": "Michel John"
                                    },
                                    "titre_document": {
                                        "type": "string",
                                        "example": "Acte de Naissance"
                                    },
                                    "date_expiration": {
                                        "type": "string",
                                        "example": "2022-08-24"
                                    },
                                    "nom_trouveur": {
                                        "type": "string",
                                        "example": "Jean Dupont"
                                    },
                                    "tel_trouveur": {
                                        "type": "string",
                                        "example": "+237693899890"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Déclaration enregistrée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Docmaster"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Déclaration créée avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/declaration/{id}": {
            "get": {
                "tags": [
                    "Déclarations"
                ],
                "summary": "Afficher une déclaration",
                "description": "Nécessite un token Bearer dans l'en-tête Authorization",
                "operationId": "8360777987c5a7405d0707df28b8a0bc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Déclaration récupérée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Docmaster"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Déclaration recupérée avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Déclarations"
                ],
                "summary": "Archiver une déclaration",
                "description": "Nécessite un token Bearer dans l'en-tête Authorization",
                "operationId": "c682e7f6bdd4239f92f03509f99e3902",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Déclaration archivée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Déclaration archivée avec succès."
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/declaration/{id}/restore": {
            "post": {
                "tags": [
                    "Déclarations"
                ],
                "summary": "Restaurer une déclaration",
                "description": "Nécessite un token Bearer dans l'en-tête Authorization",
                "operationId": "49ca9437abeeee8832cb928c830c9fe3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Déclaration restaurée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Déclaration restaurée avec succès."
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/declaration/{id}/force-delete": {
            "delete": {
                "tags": [
                    "Déclarations"
                ],
                "summary": "Suppression définitive d'une déclaration",
                "description": "Nécessite un token Bearer dans l'en-tête Authorization",
                "operationId": "96e359636b772840ab6921247bddeee1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Déclaration supprimée définitivement",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Déclaration supprimée définitivement avec succès."
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/declaration/archived": {
            "get": {
                "tags": [
                    "Déclarations"
                ],
                "summary": "Lister les déclarations archivées",
                "description": "Nécessite un token Bearer dans l'en-tête Authorization",
                "operationId": "73e7c31ffd7b43b2726a708ce01f368f",
                "responses": {
                    "200": {
                        "description": "Liste des déclarations archivées",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "archived_declarations": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Docmaster"
                                                    }
                                                },
                                                "meta": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Liste des déclarations supprimées récupérée avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/declaration/search": {
            "get": {
                "tags": [
                    "Déclarations"
                ],
                "summary": "Rechercher une déclaration par titre",
                "description": "Nécessite un token Bearer dans l'en-tête Authorization",
                "operationId": "0f09f2502ada5ca670c9e06b6d7be3cf",
                "parameters": [
                    {
                        "name": "titre",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Résultats de la recherche",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "docmasters": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Docmaster"
                                                    }
                                                },
                                                "meta": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Liste des déclarations récupérée avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Docmaster": {
                "required": [
                    "id",
                    "document_id",
                    "type_docmaster",
                    "etat_docmaster"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "uuid-docmaster-1"
                    },
                    "doc_chercheur_id": {
                        "type": "string",
                        "example": "uuid-user-1"
                    },
                    "doc_trouveur_id": {
                        "type": "string",
                        "example": "uuid-user-2"
                    },
                    "document_id": {
                        "type": "string",
                        "example": "uuid-doc"
                    },
                    "type_docmaster": {
                        "type": "string",
                        "example": "Trouver"
                    },
                    "date_action": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-06-01"
                    },
                    "nom_trouveur": {
                        "type": "string",
                        "example": "Jean Michel"
                    },
                    "infos_docs": {
                        "type": "string",
                        "example": "Trouvé au carrefour Emia"
                    },
                    "tel_trouveur": {
                        "type": "string",
                        "example": "237699800320"
                    },
                    "etat_docmaster": {
                        "type": "string",
                        "example": "Trouvé"
                    },
                    "nombre_notif": {
                        "type": "integer",
                        "example": 1
                    },
                    "credit": {
                        "type": "integer",
                        "example": 100
                    },
                    "debit": {
                        "type": "integer",
                        "example": 0
                    },
                    "confirm": {
                        "type": "boolean",
                        "example": false
                    },
                    "code_confirm": {
                        "type": "string",
                        "example": "XYZ123"
                    },
                    "supprime": {
                        "type": "boolean",
                        "example": false
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}